feat(Login): refactor login page
