--- packages/hedgehog-lab/src/components/Base/Input/Copy/ICopyInputProps.ts
@@ -1,3 +1,0 @@
-export interface ICopyInputProps {
-    url: string,
-}
